move d2-event-info in prep for DestinyItemManager/d2-additional-info/#34
