+++ src/data/d2/d2-event-info.ts
